revert spell changes
