+++ common/src/main/java/com/wynntils/core/components/Handlers.java
@@ -13,7 +13,6 @@ import com.wynntils.handlers.item.ItemHandler;
 import com.wynntils.handlers.labels.LabelHandler;
 import com.wynntils.handlers.particle.ParticleHandler;
 import com.wynntils.handlers.scoreboard.ScoreboardHandler;
-import com.wynntils.handlers.spell.SpellCastHandler;
 import com.wynntils.handlers.tooltip.TooltipHandler;
 import com.wynntils.handlers.wrappedscreen.WrappedScreenHandler;
 
@@ -27,7 +26,6 @@ public final class Handlers {
     public static final LabelHandler Label = new LabelHandler();
     public static final ParticleHandler Particle = new ParticleHandler();
     public static final ScoreboardHandler Scoreboard = new ScoreboardHandler();
-    public static final SpellCastHandler SpellCast = new SpellCastHandler();
     public static final TooltipHandler Tooltip = new TooltipHandler();
     public static final WrappedScreenHandler WrappedScreen = new WrappedScreenHandler();
 }
