+++ common/src/main/java/com/wynntils/models/spells/SpellModel.java
@@ -22,8 +22,10 @@ import com.wynntils.models.worlds.event.WorldStateEvent;
 import com.wynntils.utils.mc.McUtils;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.LinkedList;
 import java.util.List;
 import java.util.Optional;
+import java.util.Queue;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import net.neoforged.bus.api.SubscribeEvent;
@@ -33,6 +35,8 @@ public final class SpellModel extends Model {
             Pattern.compile("^§7(.*) spell cast! §3\\[§b-([0-9]+) ✺§3\\](?: §4\\[§c-([0-9]+) ❤§4\\])?$");
     private static final int SPELL_COST_RESET_TICKS = 60;
     private static final int SPELL_EXPIRE_TICKS = 40;
+
+    private static final Queue<SpellDirection> SPELL_PACKET_QUEUE = new LinkedList<>();
 
     private SpellDirection[] lastSpell = SpellDirection.NO_SPELL;
     private String lastBurstSpellName = "";
@@ -128,6 +132,7 @@ public final class SpellModel extends Model {
 
     @SubscribeEvent
     public void onWorldStateChange(WorldStateEvent e) {
+        SPELL_PACKET_QUEUE.clear();
         lastSpell = SpellDirection.NO_SPELL;
         lastSpellTick = 0;
         lastBurstSpellName = "";
@@ -140,8 +145,30 @@ public final class SpellModel extends Model {
 
     @SubscribeEvent
     public void onHeldItemChange(ChangeCarriedItemEvent event) {
+        SPELL_PACKET_QUEUE.clear();
         lastSpell = SpellDirection.NO_SPELL;
         lastSpellTick = 0;
+    }
+
+    public void addSpellToQueue(List<SpellDirection> spell) {
+        if (!SPELL_PACKET_QUEUE.isEmpty()) return;
+
+        SPELL_PACKET_QUEUE.addAll(spell);
+    }
+
+    public SpellDirection checkNextSpellDirection() {
+        return SPELL_PACKET_QUEUE.peek();
+    }
+
+    public void sendNextSpell() {
+        if (SPELL_PACKET_QUEUE.isEmpty()) return;
+
+        SpellDirection spellDirection = SPELL_PACKET_QUEUE.poll();
+        spellDirection.getSendPacketRunnable().run();
+    }
+
+    public boolean isSpellQueueEmpty() {
+        return SPELL_PACKET_QUEUE.isEmpty();
     }
 
     public String getLastBurstSpellName() {
